pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jira-confluence"
version = "0.1.0"
description = "MCP server for Atlassian Jira and Confluence REST API integration"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.20.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
jira-confluence-mcp = "jira_confluence_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "responses>=0.23.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["jira_confluence_mcp*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
